Fixed issue where auto timesheets were not created because id was larger than Int32.

+++ SSW.TimePRO.AutoTimeSheeting.Infrastructure/GitHub/GitHubModel.cs
@@ -21,7 +21,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
 
     public class Actor
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string login { get; set; }
         public string display_login { get; set; }
         public string gravatar_id { get; set; }
@@ -31,7 +31,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
 
     public class Repo
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string name { get; set; }
         public string url { get; set; }
     }
@@ -59,7 +59,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
     public class Pull_Request
     {
         public string url { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string html_url { get; set; }
         public string diff_url { get; set; }
@@ -108,7 +108,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
     public class User
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -139,7 +139,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
     public class User1
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -160,7 +160,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
 
     public class Repo1
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string name { get; set; }
         public string full_name { get; set; }
@@ -238,7 +238,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
     public class Owner
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -269,7 +269,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
     public class User2
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -290,7 +290,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
 
     public class Repo2
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string name { get; set; }
         public string full_name { get; set; }
@@ -368,7 +368,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
     public class Owner1
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -442,7 +442,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
     public class Requested_Reviewers
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -464,7 +464,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
     public class Member
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -491,7 +491,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
         public string comments_url { get; set; }
         public string events_url { get; set; }
         public string html_url { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public int number { get; set; }
         public string title { get; set; }
@@ -513,7 +513,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
     public class User3
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -534,7 +534,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
 
     public class Label
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string url { get; set; }
         public string name { get; set; }
@@ -559,7 +559,7 @@ namespace SSW.TimePRO.AutoTimeSheeting.Infrastructure.GitHub
 
     public class Org
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string login { get; set; }
         public string gravatar_id { get; set; }
         public string url { get; set; }
